# docs/swagger.yaml
basePath: /
definitions:
  controller.Info:
    properties:
      message:
        example: 消息
        type: string
      ok:
        example: true
        type: boolean
    type: object
  controller.JSONResult:
    properties:
      code:
        example: 0
        type: integer
      data:
        type: object
      info:
        $ref: '#/definitions/controller.Info'
    type: object
info:
  contact: {}
  description: 网页版kubectl cp
  title: KubeCp Swagger
  version: "1.0"
paths:
  /api/k8s/deployment:
    get:
      description: 命名空间下Deployment资源列表
      parameters:
      - in: query
        name: deployment
        type: string
      - in: query
        name: field_selector
        type: string
      - in: query
        name: label_selector
        type: string
      - in: query
        name: namespace
        type: string
      - in: query
        name: deployment
        type: string
      - in: query
        name: field_selector
        type: string
      - in: query
        name: label_selector
        type: string
      - in: query
        name: namespace
        type: string
      - in: query
        name: deployment
        type: string
      - in: query
        name: field_selector
        type: string
      - in: query
        name: label_selector
        type: string
      - in: query
        name: namespace
        type: string
      - in: query
        name: deployment
        type: string
      - in: query
        name: field_selector
        type: string
      - in: query
        name: label_selector
        type: string
      - in: query
        name: namespace
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.JSONResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.JSONResult'
      summary: ListNamespaceAllDeployment
      tags:
      - Kubernetes
  /api/k8s/download:
    get:
      consumes:
      - application/json
      description: 从容器下载到本地
      parameters:
      - in: query
        name: container_name
        type: string
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: dest_path
        required: true
        type: array
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pod_name
        required: true
        type: string
      - in: query
        name: container_name
        type: string
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: dest_path
        required: true
        type: array
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pod_name
        required: true
        type: string
      - in: query
        name: container_name
        type: string
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: dest_path
        required: true
        type: array
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pod_name
        required: true
        type: string
      - in: query
        name: container_name
        type: string
      - collectionFormat: multi
        in: query
        items:
          type: string
        name: dest_path
        required: true
        type: array
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pod_name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.JSONResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.JSONResult'
      summary: Copy2Local
      tags:
      - Kubernetes
...
  /api/k8s/multi_upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传到容器
      parameters:
      - in: query
        name: dest_path
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pod_name
        required: true
        type: string
      - in: query
        name: dest_path
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pod_name
        required: true
        type: string
      - in: query
        name: dest_path
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pod_name
        required: true
        type: string
      - description: files
        in: formData
        name: files
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.JSONResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.JSONResult'
      summary: MultiCopy2Container
      tags:
      - Kubernetes
...
  /api/k8s/upload:
    post:
      consumes:
      - multipart/form-data
      description: 上传到容器
      parameters:
      - in: query
        name: container_name
        type: string
      - in: query
        name: dest_path
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pod_name
        required: true
        type: string
      - in: query
        name: container_name
        type: string
      - in: query
        name: dest_path
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pod_name
        required: true
        type: string
      - in: query
        name: container_name
        type: string
      - in: query
        name: dest_path
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pod_name
        required: true
        type: string
      - in: query
        name: container_name
        type: string
      - in: query
        name: dest_path
        required: true
        type: string
      - in: query
        name: namespace
        required: true
        type: string
      - in: query
        name: pod_name
        required: true
        type: string
      - description: files
        in: formData
        name: files
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.JSONResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.JSONResult'
      summary: Copy2Container
      tags:
      - Kubernetes
swagger: "2.0"